leak free expand vars in heredoc
